requirements.txt
fastapi>=0.100.0
uvicorn>=0.22.0
python-jose>=3.3.0
python-multipart>=0.0.6
sqlalchemy>=2.0.0
pydantic>=2.0.0
pydantic-settings>=2.0.0
aiohttp>=3.8.5
discord.py>=2.3.0
python-dotenv>=1.0.0
passlib>=1.7.4
bcrypt>=4.0.1
aiosqlite>=0.21.0
alembic>=1.15.1